Move areSettingsVisible to the feature-toggles module
