+++ src/utils/feature-toggles.ts
@@ -1,0 +1,5 @@
+import { Env } from "src/domain";
+
+export const areSettingsVisible = (env: Env): boolean => {
+  return env.fiatExchangeRates.areEnabled;
+};
